test: extract model fallback receipt integrity cases
